requirements.txt
Flask>=1.1.1
hvac>=0.9.5
PyYAML>=5.1.2
requests-oauthlib>=1.2.0
